+++ undirected.go
@@ -58,8 +58,8 @@ func (u *undirected[K, T]) AddEdge(sourceHash, targetHash K, options ...func(*Ed
 		return fmt.Errorf("an edge between vertices %v and %v already exists", sourceHash, targetHash)
 	}
 
-	// If the graph was declared to be acyclic, permit the creation of a cycle.
-	if u.traits.IsAcyclic {
+	// If the user opted in to permitting cycles, run a cycle check.
+	if u.traits.PermitCycles {
 		createsCycle, err := CreatesCycle[K, T](u, sourceHash, targetHash)
 		if err != nil {
 			return fmt.Errorf("failed to check for cycles: %w", err)
@@ -145,6 +145,43 @@ func (u *undirected[K, T]) AdjacencyMap() (map[K]map[K]Edge[K], error) {
 
 func (u *undirected[K, T]) PredecessorMap() (map[K]map[K]Edge[K], error) {
 	return u.AdjacencyMap()
+}
+
+func (u *undirected[K, T]) Clone() (Graph[K, T], error) {
+	traits := &Traits{
+		IsDirected: u.traits.IsDirected,
+		IsAcyclic:  u.traits.IsAcyclic,
+		IsWeighted: u.traits.IsWeighted,
+		IsRooted:   u.traits.IsRooted,
+	}
+
+	vertices := make(map[K]T)
+
+	for hash, vertex := range u.vertices {
+		vertices[hash] = vertex
+	}
+
+	return &undirected[K, T]{
+		hash:     u.hash,
+		traits:   traits,
+		vertices: vertices,
+		outEdges: cloneEdges(u.outEdges),
+		inEdges:  cloneEdges(u.inEdges),
+	}, nil
+}
+
+func (u *undirected[K, T]) Order() int {
+	return len(u.vertices)
+}
+
+func (u *undirected[K, T]) Size() int {
+	size := 0
+	for _, outEdges := range u.outEdges {
+		size += len(outEdges)
+	}
+
+	// Divide by 2 since every add edge operation on undirected graph is counted twice.
+	return size / 2
 }
 
 func (u *undirected[K, T]) edgesAreEqual(a, b Edge[T]) bool {