+++ undirected_test.go
@@ -89,7 +89,7 @@ func TestUndirected_AddEdge(t *testing.T) {
 			vertices:   []int{1, 2, 3},
 			edgeHashes: [][3]int{{1, 2, 0}, {2, 3, 0}, {3, 1, 0}},
 			traits: &Traits{
-				IsAcyclic: true,
+				PermitCycles: true,
 			},
 			shouldFinallyFail: true,
 		},
@@ -413,6 +413,153 @@ func TestUndirected_PredecessorMap(t *testing.T) {
 				}
 			}
 		}
+	}
+}
+
+func TestUndirected_Clone(t *testing.T) {
+	tests := map[string]struct {
+		vertices []int
+		edges    []Edge[int]
+	}{
+		"Y-shaped graph": {
+			vertices: []int{1, 2, 3, 4},
+			edges: []Edge[int]{
+				{Source: 1, Target: 3},
+				{Source: 2, Target: 3},
+				{Source: 3, Target: 4},
+			},
+		},
+		"diamond-shaped graph": {
+			vertices: []int{1, 2, 3, 4},
+			edges: []Edge[int]{
+				{Source: 1, Target: 2},
+				{Source: 1, Target: 3},
+				{Source: 2, Target: 4},
+				{Source: 3, Target: 4},
+			},
+		},
+	}
+
+	for name, test := range tests {
+		graph := New(IntHash)
+
+		for _, vertex := range test.vertices {
+			_ = graph.AddVertex(vertex)
+		}
+
+		for _, edge := range test.edges {
+			if err := graph.AddEdge(edge.Source, edge.Target, EdgeWeight(edge.Properties.Weight)); err != nil {
+				t.Fatalf("%s: failed to add edge: %s", name, err.Error())
+			}
+		}
+
+		clonedGraph, err := graph.Clone()
+		if err != nil {
+			t.Fatalf("%s: failed to clone graph: %s", name, err.Error())
+		}
+
+		expected := graph.(*undirected[int, int])
+		actual := clonedGraph.(*undirected[int, int])
+
+		if actual.hash(5) != expected.hash(5) {
+			t.Errorf("%s: hash expectancy doesn't match: expected %v, got %v", name, expected.hash, actual.hash)
+		}
+
+		if !traitsAreEqual(actual.traits, expected.traits) {
+			t.Errorf("%s: traits expectancy doesn't match: expected %v, got %v", name, expected.traits, actual.traits)
+		}
+
+		if len(actual.vertices) != len(expected.vertices) {
+			t.Fatalf("%s: vertices length expectancy doesn't match: expected %v, got %v", name, len(expected.vertices), len(actual.vertices))
+		}
+
+		for expectedHash, expectedVertex := range expected.vertices {
+			actualVertex, ok := actual.vertices[expectedHash]
+			if !ok {
+				t.Errorf("%s: vertex expectancy doesn't match: expected vertex %v doesn't exist", name, expectedVertex)
+			}
+			if actualVertex != expectedVertex {
+				t.Errorf("%s: vertex expectancy doesn't match: expected %v, got %v", name, expectedVertex, actualVertex)
+			}
+		}
+
+		if len(actual.inEdges) != len(expected.inEdges) {
+			t.Errorf("%s: number of inEdges doesn't match: expected %v, got %v", name, len(expected.inEdges), len(actual.inEdges))
+		}
+		if len(actual.outEdges) != len(expected.outEdges) {
+			t.Errorf("%s: number of outEdges doesn't match: expected %v, got %v", name, len(expected.outEdges), len(actual.outEdges))
+		}
+	}
+}
+
+func TestUndirected_OrderAndSize(t *testing.T) {
+	tests := map[string]struct {
+		vertices      []int
+		edges         []Edge[int]
+		expectedOrder int
+		expectedSize  int
+	}{
+		"Y-shaped graph": {
+			vertices: []int{1, 2, 3, 4},
+			edges: []Edge[int]{
+				{Source: 1, Target: 3},
+				{Source: 2, Target: 3},
+				{Source: 3, Target: 4},
+			},
+			expectedOrder: 4,
+			expectedSize:  3,
+		},
+		"diamond-shaped graph": {
+			vertices: []int{1, 2, 3, 4},
+			edges: []Edge[int]{
+				{Source: 1, Target: 2},
+				{Source: 1, Target: 3},
+				{Source: 2, Target: 4},
+				{Source: 3, Target: 4},
+			},
+			expectedOrder: 4,
+			expectedSize:  4,
+		},
+		"two-vertices graph": {
+			vertices: []int{1, 2},
+			edges: []Edge[int]{
+				{Source: 1, Target: 2},
+			},
+			expectedOrder: 2,
+			expectedSize:  1,
+		},
+		"edgeless graph": {
+			vertices:      []int{1, 2},
+			edges:         []Edge[int]{},
+			expectedOrder: 2,
+			expectedSize:  0,
+		},
+	}
+
+	for name, test := range tests {
+		graph := newUndirected(IntHash, &Traits{})
+
+		for _, vertex := range test.vertices {
+			_ = graph.AddVertex(vertex)
+		}
+
+		for _, edge := range test.edges {
+			if err := graph.AddEdge(edge.Source, edge.Target, EdgeWeight(edge.Properties.Weight)); err != nil {
+				t.Fatalf("%s: failed to add edge: %s", name, err.Error())
+			}
+		}
+
+		order := graph.Order()
+		size := graph.Size()
+
+		if order != test.expectedOrder {
+			t.Errorf("%s: order expectancy doesn't match: expected %d, got %d", name, test.expectedOrder, order)
+		}
+
+		if size != test.expectedSize {
+			t.Errorf("%s: size expectancy doesn't match: expected %d, got %d", name, test.expectedSize, size)
+		}
+
 	}
 }